Refactor transition state storage to use uint256
Changed TransitionState.batchIdAndPartialParentHash from bytes32 to uint256 for more efficient storage and manipulation. Inlined batchId and partialParentHash encoding/decoding logic into TaikoInbox2, removed related helpers from LibData2, and updated function signatures and usages accordingly.

+++ packages/protocol/contracts/layer1/based2/libs/LibProve2.sol
@@ -28,7 +28,7 @@ library LibProve2 {
         // writes
         function(address, uint256) creditBond;
         function(I.Config memory, address, uint256) debitBond;
-        function(I.Config memory, uint256, bytes32, bytes32) returns (bool) saveTransition;
+        function(I.Config memory, uint48, bytes32, bytes32) returns (bool) saveTransition;
     }
 
     function proveBatches(
